# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API Documentation for your application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/beneficios/{user_id}": {
            "get": {
                "tags": [
                    "Beneficios"
                ],
                "summary": "Obtener los beneficios de un usuario",
                "operationId": "c70063c1634e06639e6d6d38cf2d5a61",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Beneficios del usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ID del beneficio",
                                                "type": "integer"
                                            },
                                            "monto": {
                                                "description": "Monto del beneficio",
                                                "type": "integer"
                                            },
                                            "fecha_recepcion": {
                                                "description": "Fecha de recepción del beneficio",
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "fecha": {
                                                "description": "Fecha del beneficio",
                                                "type": "string",
                                                "format": "date"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            }
        }
    }
}